refacto: move HIR analysis into new module
